update concerns dir structuring for autoloading purposes
